Started EvolutionaryAlgorithm class

+++ src/main/java/com/oast/Module.java
@@ -1,17 +1,17 @@
 package com.oast;
 
 public class Module {
-    private float cost;
+    private double cost;
 
-    public Module(float cost) {
+    public Module(double cost) {
         this.cost = cost;
     }
 
-    public float getCost() {
+    public double getCost() {
         return cost;
     }
 
-    public void setCost(float cost) {
+    public void setCost(double cost) {
         this.cost = cost;
     }
 }
